requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Flask-Babel==3.1.0
Flask-Limiter==3.5.0

# Database
SQLAlchemy==2.0.21
Alembic==1.12.0
PyMySQL==1.1.0
psycopg2-binary==2.9.7

# Forms and validation
WTForms==3.0.1
email-validator==2.0.0

# Security
Werkzeug==2.3.7
bcrypt==4.0.1

# Image processing
Pillow==10.0.1

# Utilities
python-dotenv==1.0.0
click==8.1.7
Jinja2==3.1.2

# Production server
gunicorn==21.2.0

# Development tools
pytest==7.4.2
pytest-flask==1.2.0
coverage==7.3.2
